pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "route-analyzer"
version = "0.1.0"
description = "Route decision analysis toolkit for trajectory data"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0", 
    "matplotlib>=3.3.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.60.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
yaml = ["pyyaml"]
parquet = ["pyarrow"]
gui = ["streamlit>=1.28.0", "plotly>=5.15.0"]
test = ["pytest>=7.0.0"]

[project.scripts]
route-analyzer = "route_analyzer.__main__:main"

[tool.setuptools]
packages = ["route_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
pythonpath = ["."]